build script is now checking if libs/Native exists before injecting native dependencies;

+++ build.fsx
@@ -95,39 +95,39 @@ Target "CreatePackage" (fun () ->
 
 
 Target "InjectNativeDependencies" (fun () ->
+    if Directory.Exists "libs/Native" then
+        if Directory.Exists "bin/Debug" then
+            File.Copy(@"packages/Aardvark.Build/lib/net45/Aardvark.Build.dll", @"bin/Debug/Aardvark.Build.dll", true)
+        if Directory.Exists "bin/Release" then
+            File.Copy(@"packages/Aardvark.Build/lib/net45/Aardvark.Build.dll", @"bin/Release/Aardvark.Build.dll", true)
 
-    if Directory.Exists "bin/Debug" then
-        File.Copy(@"packages/Aardvark.Build/lib/net45/Aardvark.Build.dll", @"bin/Debug/Aardvark.Build.dll", true)
-    if Directory.Exists "bin/Release" then
-        File.Copy(@"packages/Aardvark.Build/lib/net45/Aardvark.Build.dll", @"bin/Release/Aardvark.Build.dll", true)
+        let dirs = Directory.GetDirectories "libs/Native"
+        for d in dirs do
+            let n = Path.GetFileName d
+            let d = d |> Path.GetFullPath
+            let paths = [
+                Path.Combine("bin/Release", n + ".dll") |> Path.GetFullPath
+                Path.Combine("bin/Release", n + ".exe") |> Path.GetFullPath
+                Path.Combine("bin/Debug", n + ".dll") |> Path.GetFullPath
+                Path.Combine("bin/Debug", n + ".exe") |> Path.GetFullPath
+            ]
 
-    let dirs = Directory.GetDirectories "libs/Native"
-    for d in dirs do
-        let n = Path.GetFileName d
-        let d = d |> Path.GetFullPath
-        let paths = [
-            Path.Combine("bin/Release", n + ".dll") |> Path.GetFullPath
-            Path.Combine("bin/Release", n + ".exe") |> Path.GetFullPath
-            Path.Combine("bin/Debug", n + ".dll") |> Path.GetFullPath
-            Path.Combine("bin/Debug", n + ".exe") |> Path.GetFullPath
-        ]
-
-        let wd = Environment.CurrentDirectory
-        try
+            let wd = Environment.CurrentDirectory
+            try
         
-            for p in paths do
-                if File.Exists p then
-                    Environment.CurrentDirectory <- Path.GetDirectoryName p
-                    let ass = Mono.Cecil.AssemblyDefinition.ReadAssembly(p)
-                    AssemblyInjector.addResources d ass
-                    ass.Write p
-                    tracefn "injected native stuff in %A" p
-        finally
-            Environment.CurrentDirectory <- wd
+                for p in paths do
+                    if File.Exists p then
+                        Environment.CurrentDirectory <- Path.GetDirectoryName p
+                        let ass = Mono.Cecil.AssemblyDefinition.ReadAssembly(p)
+                        AssemblyInjector.addResources d ass
+                        ass.Write p
+                        tracefn "injected native stuff in %A" p
+            finally
+                Environment.CurrentDirectory <- wd
+
+            ()
 
         ()
-
-    ()
 )
 
 Target "Push" (fun () ->
